feat: Allow users to set display names
Introduces functionality for users to set and save a display name, which will be shown to other users. This display name is now included in the incoming call payload to inform the recipient of who is calling.

Additionally, the connection status indicator has been refined to more clearly display the encrypted state with an icon.